create code/manager
