+++ code/manager/package.json
@@ -1,0 +1,11 @@
+{
+  "name": "@storybook/manager",
+  "private": true,
+  "type": "module",
+  "packageManager": "yarn@3.2.3",
+  "dependencies": {
+    "@vitejs/plugin-react": "^2.1.0",
+    "typescript": "^4.8.4",
+    "vite": "^3.1.7"
+  }
+}
