+++ code/package.json
@@ -41,13 +41,15 @@
     "packages": [
       "addons/*",
       "addons/storyshots/*",
-      "frameworks/*",
-      "renderers/*",
-      "presets/*",
-      "examples-native/*",
       "examples/*",
+      "examples-native/*",
+      "frameworks/*",
       "lib/*",
-      "lib/cli/test/run/*"
+      "lib/cli/test/run/*",
+      "manager",
+      "manager/*",
+      "presets/*",
+      "renderers/*"
     ]
   },
   "scripts": {
@@ -80,7 +82,7 @@
     "sandbox": "ts-node ../scripts/sandbox.ts",
     "serve-storybooks": "http-server ./built-storybooks -p 8001",
     "smoketest-storybooks": "cross-env STORYBOOK_DISPLAY_WARNING=true DISPLAY_WARNING=true node -r esm ../scripts/smoketest-storybooks.js",
-    "storybook:ui": "./lib/cli/bin/index.js dev --config-dir sb-for-sb-ui --port 6006 --no-manager-cache",
+    "storybook:ui": "cd manager && ../lib/cli/bin/index.js dev --port 6006 --no-manager-cache",
     "test": "NODE_OPTIONS=--max_old_space_size=4096 jest --config ./jest.config.js",
     "test:cli": "npm --prefix lib/cli run test",
     "test:e2e-examples-playwright": "playwright test",
